changes build_training_sample to staticmethod
