+++ mantisshrimp/models/mantis_rcnn/__init__.py
@@ -1,5 +1,4 @@
 from .rcnn_param_groups import *
-from .build_training_sample_rcnn_mixin import *
 from .mantis_rcnn import *
 from .mantis_faster_rcnn import *
 from .mantis_mask_rcnn import *
