+++ tests/models/test_rcnn_build_training_sample.py
@@ -5,7 +5,7 @@ _fake_box = [0, 1, 2, 3]
 
 
 def test_build_training_sample_maskrcnn(data_sample):
-    x, y = BuildTrainingSampleMaskRCNNMixin().build_training_sample(**data_sample)
+    x, y = MantisMaskRCNN.build_training_sample(**data_sample)
     assert x.dtype == torch.float32
     assert x.shape == (3, 427, 640)
     assert isinstance(y, dict)
@@ -30,7 +30,7 @@ def test_build_training_sample_maskrcnn(data_sample):
 
 
 def test_build_training_sample_maskrcnn(data_sample):
-    x, y = BuildTrainingSampleFasterRCNNMixin().build_training_sample(**data_sample)
+    x, y = MantisFasterRCNN.build_training_sample(**data_sample)
     assert x.dtype == torch.float32
     assert x.shape == (3, 427, 640)
     assert isinstance(y, dict)
@@ -50,7 +50,7 @@ def test_rcnn_empty_training_sample(data_sample):
     data_sample["bbox"] = []
     data_sample["label"] = []
     data_sample["iscrowd"] = []
-    x, y = BuildTrainingSampleMaskRCNNMixin().build_training_sample(**data_sample)
+    x, y = MantisMaskRCNN.build_training_sample(**data_sample)
     assert (y["boxes"] == tensor([_fake_box], dtype=torch.float)).all()
     assert y["labels"] == tensor([0])
     assert y["iscrowd"] == tensor([0])
